LogstashAgent/pyproject.toml
[build-system]
requires = ["uv_build>=0.11.2,<0.12"]
build-backend = "uv_build"

[project]
name = "LogstashAgent"
version = "0.1.0"
description = "An agent that is used to entirely control an instance of Logstash"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asn1crypto>=1.5.1",
    "cryptography>=46.0.6",
    "fastapi>=0.135.2",
    "httpx>=0.28.1",
    "psutil>=7.2.2",
    "pytest>=9.0.2",
    "pyyaml>=6.0.3",
    "requests>=2.33.0",
    "uvicorn>=0.42.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2,<10",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]

[project.scripts]
logstash-agent = "logstashagent.main:main"